Fix ct cover specs detection
If there are project deps with their own ct cover specs, rebar will
currently incorrectly detect all of them and fail, since the code for
removing the deps does not work.

This patch fixes the code by breaking the deps path up and removing the
common prefix for deps.  This may fail if deps is a director in the base
directory, since it expects a single item in a list.